requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy>=2.0
psycopg[binary]
alembic
pydantic-settings
bcrypt>=4.0
PyJWT>=2.8
